// src/shared/types/generated/application/SidebarCounts.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ScopeCount } from "./ScopeCount";

export type SidebarCounts = { all: number, inbox: number, trash: number, recently_viewed: number, untagged: number, uncategorized: number, duplicates: number, folders: Array<ScopeCount>, smart_folders: Array<ScopeCount>, revision: number, };